# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: AnchorPay Public API
  version: 1.0.0
  summary: Customer, staff and webhook API exposed through the API gateway.
  description: |
    All client traffic enters through the API gateway, which validates the JWT, applies rate limits,
    adds `X-Request-Id` and routes each path to the service named in `x-owner-service`.

    **Conventions** (frozen at Foundation sign-off)
    - JSON bodies use camelCase. Money is `{ amountMinor, currency }` in integer minor units; rates are decimal strings.
    - Every error is `application/problem+json` with a stable `code` (see `ErrorCode`).
    - `x-roles` lists the roles allowed to call an operation. Customers only ever see their own resources.
    - Versioning: breaking changes go to `/v2`; additive changes (new optional fields, new endpoints) stay in `/v1`.
    - Access tokens expire after 15 minutes; refresh tokens after 30 days (rotated on every refresh).
  contact:
    name: AnchorPay engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local API gateway
security:
  - bearerAuth: []
...
paths:
...
  /v1/users/me:
    get:
      tags: [Profile]
      operationId: getMe
      summary: Get the signed-in user's profile
      x-owner-service: identity-service
      x-roles: [customer, agent, compliance_officer, admin]
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
    patch:
      tags: [Profile]
      operationId: updateMe
      summary: Update phone or address
      description: >-
        Name and date of birth can't be changed here because they are tied to identity verification (contact support).
        Changing the phone number requires verifying the new number.
      x-owner-service: identity-service
      x-roles: [customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '409':
          $ref: 'common.yaml#/components/responses/Conflict'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/users/me/change-password:
    post:
      tags: [Profile]
      operationId: changePassword
      summary: Change password (requires the current password)
      x-owner-service: identity-service
      x-roles: [customer, agent, compliance_officer, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  $ref: '#/components/schemas/Password'
      responses:
        '204':
          description: Changed; other sessions revoked
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'

  # ---------------------------------------------------------------- recipients
  /v1/recipients:
    get:
      tags: [Recipients]
      operationId: listRecipients
      summary: List saved recipients
      x-owner-service: identity-service
      x-roles: [customer]
      responses:
        '200':
          description: Recipients (newest first)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipient'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
    post:
      tags: [Recipients]
      operationId: createRecipient
      summary: Save a new recipient
      x-owner-service: identity-service
      x-roles: [customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientCreate'
      responses:
        '201':
          description: Recipient saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '422':
          $ref: 'common.yaml#/components/responses/UnprocessableEntity'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
  /v1/recipients/{recipientId}:
    parameters:
      - $ref: '#/components/parameters/RecipientId'
    get:
      tags: [Recipients]
      operationId: getRecipient
      summary: Get one recipient
      x-owner-service: identity-service
      x-roles: [customer]
      responses:
        '200':
          description: Recipient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
    patch:
      tags: [Recipients]
      operationId: updateRecipient
      summary: Update nickname, relationship or phone
      description: Payout details (bank account / wallet) are immutable — save a new recipient instead.
      x-owner-service: identity-service
      x-roles: [customer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'
    delete:
      tags: [Recipients]
      operationId: deleteRecipient
      summary: Remove a recipient from the address book (soft delete; past transfers keep it)
      x-owner-service: identity-service
      x-roles: [customer]
      responses:
        '204':
          description: Removed
        '401':
          $ref: 'common.yaml#/components/responses/Unauthorized'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
        default:
          $ref: 'common.yaml#/components/responses/ServerError'

  # ---------------------------------------------------------------- pricing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'RS256 access token from /v1/auth/login. Claims: sub (user id), role, sid (session id), iss, aud, exp (15 min). KYC tier is not a claim: it can change at any time, ask compliance-service.'
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
    providerSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: 'Hex HMAC-SHA256 of "<X-Timestamp>.<raw body>" with the provider webhook secret. Rejected if older than 5 minutes.'
  parameters:
...
    RecipientId:
      name: recipientId
      in: path
      required: true
      schema:
        $ref: 'common.yaml#/components/schemas/Uuid'
...
  schemas:
...
    Password:
      type: string
      minLength: 12
      maxLength: 128
      writeOnly: true
      description: At least 12 characters. Checked against a list of common/breached passwords.
    Address:
      type: object
      required: [line1, city, province, postalCode]
      properties:
        line1:
          type: string
          maxLength: 200
        line2:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        province:
          type: string
          enum: [AB, BC, MB, NB, NL, NS, NT, NU, 'ON', PE, QC, SK, YT]
        postalCode:
          type: string
          pattern: '^[A-Za-z][0-9][A-Za-z] ?[0-9][A-Za-z][0-9]$'
...
    User:
      type: object
      required: [id, email, phone, fullName, country, role, status, emailVerified, phoneVerified, createdAt]
      properties:
        id:
          $ref: 'common.yaml#/components/schemas/Uuid'
        email:
          type: string
          format: email
        phone:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
        country:
          $ref: 'common.yaml#/components/schemas/Country'
        role:
          $ref: 'common.yaml#/components/schemas/Role'
        status:
          type: string
          enum: [active, locked, suspended, closed]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        createdAt:
          $ref: 'common.yaml#/components/schemas/Timestamp'
    UpdateMeRequest:
      type: object
      minProperties: 1
      properties:
        phone:
          type: string
          pattern: '^\+[1-9][0-9]{7,14}$'
        address:
          $ref: '#/components/schemas/Address'
    Recipient:
      type: object
      required: [id, fullName, country, currency, payoutMethod, createdAt, updatedAt]
      properties:
        id:
          $ref: 'common.yaml#/components/schemas/Uuid'
        nickname:
          type: string
        fullName:
          type: string
        phone:
          type: string
        country:
          $ref: 'common.yaml#/components/schemas/Country'
        currency:
          $ref: 'common.yaml#/components/schemas/Currency'
        relationship:
          type: string
        payoutMethod:
          $ref: 'common.yaml#/components/schemas/PayoutMethod'
        bankAccount:
          type: object
          required: [bankName, accountNumberMasked]
          properties:
            bankName:
              type: string
            bankCode:
              type: string
            accountNumberMasked:
              type: string
              examples: ['•••• 4521']
        mobileWallet:
          type: object
          required: [provider, walletNumberMasked]
          properties:
            provider:
              type: string
              enum: [jazzcash, easypaisa]
            walletNumberMasked:
              type: string
              examples: ['+92 3•• ••• 4521']
        createdAt:
          $ref: 'common.yaml#/components/schemas/Timestamp'
        updatedAt:
          $ref: 'common.yaml#/components/schemas/Timestamp'
    RecipientCreate:
      type: object
      description: Provide bankAccount when payoutMethod is bank_account, mobileWallet when it is mobile_wallet.
      required: [fullName, country, currency, payoutMethod]
      properties:
        nickname:
          type: string
          maxLength: 60
        fullName:
          type: string
          minLength: 2
          maxLength: 150
        phone:
          type: string
          pattern: '^\+[1-9][0-9]{7,14}$'
        country:
          $ref: 'common.yaml#/components/schemas/Country'
        currency:
          $ref: 'common.yaml#/components/schemas/Currency'
        relationship:
          type: string
          enum: [family, friend, spouse, parent, child, sibling, business, self, other]
        payoutMethod:
          $ref: 'common.yaml#/components/schemas/PayoutMethod'
        bankAccount:
          type: object
          required: [bankName, accountNumber]
          properties:
            bankName:
              type: string
            bankCode:
              type: string
            accountNumber:
              type: string
              description: Account number or IBAN (Pakistan IBAN is 24 characters, starting PK).
              minLength: 6
              maxLength: 34
              writeOnly: true
        mobileWallet:
          type: object
          required: [provider, walletNumber]
          properties:
            provider:
              type: string
              enum: [jazzcash, easypaisa]
            walletNumber:
              type: string
              pattern: '^\+92[0-9]{10}$'
              writeOnly: true
      oneOf:
        - required: [bankAccount]
          properties:
            payoutMethod:
              const: bank_account
        - required: [mobileWallet]
          properties:
            payoutMethod:
              const: mobile_wallet
    RecipientUpdate:
      type: object
      minProperties: 1
      properties:
        nickname:
          type: string
          maxLength: 60
        relationship:
          type: string
          enum: [family, friend, spouse, parent, child, sibling, business, self, other]
        phone:
          type: string
          pattern: '^\+[1-9][0-9]{7,14}$'
...